move searching/grouping functionality from controller into hookable actions
